Fix opening link issue in Linux
In this commit I wrote a utility for opening links in a browser window, since there are mutliple locations where this happens and this implementation might need to change over time.  I also changed the message about copy and pasting the eula link, since it is currently not possible to select text on a jFX alert window.