Fix sonar code smells
I have removed the additional incorrect login dialog because it was too cumbersome to get it to work without the sun library (which should not be used because it might be dropped in future versions of java). Instead, I put a label next to the buttons on the authentication dialog which is visible when incorrect login details were used. I also changed some other minor things.