Fix issue with duplicate path parts in project selection
When you are adding a new project and browse to the folder that is already selected before specifying a name for the project, the newly selected project name will be appended to the selected project folder which is confusing. I fixed this issue by not triggering a project name change if the selected path is equal to the previously selected path.